Написал тесты, исправил баги, подправил исключения.

+++ src/main/java/ru/shipov/rest/repository/ApplicationRepository.java
@@ -6,11 +6,11 @@ import ru.shipov.rest.entity.Application;
 import java.util.List;
 import java.util.Optional;
 
-public interface ApplicationRepository extends CrudRepository<Application, Long> {
+public interface ApplicationRepository extends CrudRepository<Application, Integer> {
 
-    public Optional<Application> findById(Long contactId);
+    public Optional<Application> findById(Integer contactId);
 
     public List<Application> findAll();
 
-    public Application findFirstByContactIdOrderByDtCreatedDesc(Long contactId);
+    public Application findFirstByContactIdOrderByDtCreatedDesc(Integer contactId);
 }
